Add configuration option for monster to always follow a player position.
Adds the functionality of a per-arena setting 'monster-tracks-player' to make monsters always follow a player location. This fixes the issue of mobs wandering when they are unable to find a closest target due to it returning null. This solution creates a HashMap of monsters and their most recent player targets. An overloaded getClosestPlayer method is called if the new configuration option is true. This will return the most recent player target or a random player in the arena if the previous target died or left and it is unable to find a closest target.

This is a solution to #412.