Add configuration option for monster to glow after specified delay time.
Adds the functionality of a per-arena setting 'monster-glow-delay' to make monsters always follow a player location. This fixes the issue of spending time searching for mobs in larger arenas. This solution creates a HashMap of monsters and ties them to the wave in which they were spawned. A BukkitRunnable is created to run after the seconds defined in the monsterGlowDelay variable. This will apply a GLOWING potion effect to any monster that is still alive from the wave for which the task is called for. This method allows future use of modifying monsters on a per-wave basis and not every current living monster.

This is a solution to #629.

Co-authored-by: DDerthickJr <25013744+DDerthickJr@users.noreply.github.com>