Refactor throw and pull abilities into utility class.
Introduces the _Tosser_ utility class, which offers yeet and yoink
methods for use by the throw and pull abilities, respectively.

The primary driver for this refactoring is actually to add debug logging
into the logic to hopefully discover what the underlying bug behind the
mysterious "instant deaths" allegedly caused by the throw and/or pull
abilities. The debugging is added in a later commit to give way to a
simple revert when we no longer need it.